// MyMusic.Client/src/model/listBackgroundJobFailuresItem.ts
// @ts-nocheck
/**
Generated by orval 🍺
Do not edit manually.
MyMusic.Server | v1
OpenAPI spec version: 1.0.0
*/import type { ListBackgroundJobFailureDetail } from './listBackgroundJobFailureDetail';

export interface ListBackgroundJobFailuresItem {
  id: string;
  title: string;
  /** @nullable */
  message: string | null;
  /** @nullable */
  occurredAt: string | null;
  details: ListBackgroundJobFailureDetail[];
}